pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "balanced-spiking-networks-pc"
version = "0.1.0"
description = "Balanced spiking predictive-coding networks (Boerlin et al., 2013)."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "balanced-spiking-networks-pc contributors" }]
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "matplotlib>=3.7",
  "pandas>=2.0",
  "typer>=0.12",
  "pyyaml>=6.0",
  "pytest>=8.0",
]

[project.scripts]
bsn-pc = "bsn_pc.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/bsn_pc"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src"]
